Prevent moving overlay using arrows when it's locked.
Currently, even if the overlay is locked, you move it when you press arrows. This is unexpected because it prevents using these keys to move around the page. 

In this commit, I just add a check whether the overlay is locked and prevent the movement if it isn't.
I chose to use an additional variable for this because 1) it's more flexible - you can always add/remove locking of other keys, and 2) it doesn't mess with the current code structure.